+++ CHANGELOG.md
@@ -1308,6 +1308,10 @@ should use 4.0.1-alpha.0 for testing.
 -   `formatTransaction` will now replace `data` transaction property with `input` (#5915)
 -   `isTransactionCall` will now check if `value.input` `isHexStrict` if provided (#5915)
 
+#### web3-eth-accounts
+
+-   Moved @ethereumjs/tx, @ethereumjs/common code to our source code (#5963)
+
 #### web3-eth-contract
 
 -   `getSendTxParams` will now return `input` instead of `data` in returned transaction parameters object (#5915)
@@ -1336,7 +1340,6 @@ should use 4.0.1-alpha.0 for testing.
 
 #### web3-eth
 
--   Added hybrid build (ESM and CJS) of library (#5904)
 -   Added source files (#5956)
 
 #### web3-eth-abi
@@ -1409,12 +1412,17 @@ should use 4.0.1-alpha.0 for testing.
 
 -   Added source files (#5956)
 -   Added hybrid build (ESM and CJS) of library (#5904)
+-   Added functions `isHexString`, `isHexPrefixed`, `validateNoLeadingZeroes` (#5963)
 
 ### Removed
 
 #### web3-core
 
 -   `getConfig` method from `Web3Config` class, `config` is now public and accessible using `Web3Config.config` (#5950)
+
+#### web3-eth
+
+-   Removed dependencies @ethereumjs/tx, @ethereumjs/common (#5963)
 
 #### web3-eth-abi
 
@@ -1423,6 +1431,10 @@ should use 4.0.1-alpha.0 for testing.
 #### web3-eth-contract
 
 -   `data` was removed as a property of `ContractOptions` type (#5915)
+
+#### web3-utils
+
+-   Removed dependencies @ethereumjs/tx, @ethereumjs/common (#5963)
 
 ### Fixed