+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -77,3 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Added source files (#5956)
 -   Added hybrid build (ESM and CJS) of library (#5904)
+
+### Changed
+
+-   Moved @ethereumjs/tx, @ethereumjs/common code to our source code (#5963)